Feature/backend controller

+++ backend/Entities/Achievements/AchievementPost.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using backend.Entities.Common;
 
 namespace backend.Entities.Achievements
 {
@@ -8,11 +9,11 @@ namespace backend.Entities.Achievements
     {
         [Key]
         [Column("pk_post_id")]
-        public int PostId { get; set; }
+        public long PostId { get; set; }
 
         [Required]
         [Column("fk_author_id")]
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
 
         [Required]
         [MaxLength(255)]
