+++ backend/Entities/Referrals/JobOpening.cs
@@ -1,5 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using backend.Entities.Common;
 
 namespace backend.Entities.Referrals
 {
@@ -8,7 +9,7 @@ namespace backend.Entities.Referrals
     {
         [Key]
         [Column("pk_job_id")]
-        public int JobId { get; set; }
+        public long JobId { get; set; }
 
         [Required]
         [MaxLength(255)]
@@ -48,7 +49,7 @@ namespace backend.Entities.Referrals
         public string? CvReviewerEmails { get; set; }
 
         [Column("fk_posted_by")]
-        public int? PostedBy { get; set; }
+        public long? PostedBy { get; set; }
 
         [Column("posted_at")]
         public DateTime PostedAt { get; set; } = DateTime.UtcNow;
